seperated CPH resolvers as per suggestion
