+++ app/graphql/resolvers/business/query.js
@@ -4,7 +4,6 @@ import {
 import {
   transformOrganisationToBusiness
 } from '../../../transformers/rural-payments-portal/business.js'
-import { transformOrganisationCPHInfo } from '../../../transformers/rural-payments-portal/business-cph.js'
 
 export const Query = {
   async business (__, { id }, { dataSources }) {
@@ -22,14 +21,4 @@ export const Query = {
 
     return transformOrganisationCSApplicationToBusinessApplications(response.applications)
   }
-}
-
-export const CPH = async (_, { id }, { dataSources }) => {
-  return await dataSources.ruralPaymentsPortalApi.getOrganisationCPHCollectionBySBI(id)
-}
-
-export const CPHInfo = async ({ business, cphNumber }, _, { dataSources }) => {
-  const response = await dataSources.ruralPaymentsPortalApi.getOrganisationCPHInfoBySBIAndCPHNumber(business.id, cphNumber)
-
-  return transformOrganisationCPHInfo(response)
 }
