+++ test/integration/graphql/business.test.js
@@ -94,7 +94,7 @@ describe('Query.customer', () => {
           ...JSON.parse(JSON.stringify(transformedOrganisation)),
           cph: {
             ...JSON.parse(JSON.stringify(organisationCPHFixture)),
-            info: JSON.parse(JSON.stringify(transformedOrganisationCPHInfo)),
+            info: JSON.parse(JSON.stringify(transformedOrganisationCPHInfo))
           }
         }
       }
